+++ model_pb_trx.go
@@ -8,17 +8,18 @@
  */
 
 package rum_sdk
+
 // PbTrx struct for PbTrx
 type PbTrx struct {
-	Data []int32 `json:"Data,omitempty"`
-	Expired int32 `json:"Expired,omitempty"`
-	GroupId string `json:"GroupId,omitempty"`
-	ResendCount int32 `json:"ResendCount,omitempty"`
-	SenderPubkey string `json:"SenderPubkey,omitempty"`
-	SenderSign []int32 `json:"SenderSign,omitempty"`
-	StorageType int32 `json:"StorageType,omitempty"`
-	TimeStamp string `json:"TimeStamp,omitempty"`
-	TrxId string `json:"TrxId,omitempty"`
-	Type int32 `json:"Type,omitempty"`
-	Version string `json:"Version,omitempty"`
+	Data         []int64 `json:"Data,omitempty"`
+	Expired      int64   `json:"Expired,omitempty"`
+	GroupId      string  `json:"GroupId,omitempty"`
+	ResendCount  int64   `json:"ResendCount,omitempty"`
+	SenderPubkey string  `json:"SenderPubkey,omitempty"`
+	SenderSign   []int64 `json:"SenderSign,omitempty"`
+	StorageType  int64   `json:"StorageType,omitempty"`
+	TimeStamp    string  `json:"TimeStamp,omitempty"`
+	TrxId        string  `json:"TrxId,omitempty"`
+	Type         int64   `json:"Type,omitempty"`
+	Version      string  `json:"Version,omitempty"`
 }
